Split the attachment components into separate files.
