+++ src/assets/index.d.ts
@@ -2,3 +2,8 @@ declare module '*.png' {
     const value: any;
     export = value;
 }
+
+declare module '*.svg' {
+    const value: any;
+    export = value;
+}
