+++ src/core/attachment/index.js
@@ -1,5 +1,9 @@
 import DownloadButton from "./download-button";
+import Image from "./image-attachment";
+import Upload from "./upload-component";
 
 export {
-    DownloadButton
+    DownloadButton,
+    Image,
+    Upload
 }
