+++ src/reports/income-expense/top-account.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import AccountRepository from "../../core/repositories/account-repository";
-import { Attachments, Formats, Layout, Translations } from "../../core";
+import { Attachment, Formats, Layout, Translations } from "../../core";
 import { Account } from "../../core/types";
 
 type TopAccountTableProps = {
@@ -39,7 +39,7 @@ const TopAccountTable = ({ year, type }: TopAccountTableProps) => {
 
             { accounts && accounts.map(account => (
                 <tr key={account.account.id}>
-                    <td width={ 50 }><Attachments.Image fileCode={account.account.iconFileCode}/></td>
+                    <td width={ 50 }><Attachment.Image fileCode={account.account.iconFileCode}/></td>
                     <td>{account.account.name}</td>
                     <td><Formats.Money money={account.total * -1} currency={account.account.account.currency}/></td>
                     <td><Formats.Money money={account.average * -1} currency={account.account.account.currency}/></td>
